% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wctag.R
\docType{package}
\name{WCtag-package}
\alias{WCtag-package}
\title{R Tools for Wildlife Computer Tags}
\description{
R Tools for Wildlife Computer Tags, in support of the \pkg{SGAT} and
\pkg{tripEstimation} packages.
}
\author{
Michael D. Sumner
}

